refactor: create @md/shared package for common code
