+++ apps/web/src/components/CodemirrorEditor/EditorHeader/PostInfo.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import type { Post, PostAccount } from '@/types'
+import type { Post, PostAccount } from '@md/shared/types'
 import { Check, Info } from 'lucide-vue-next'
 import { CheckboxIndicator, CheckboxRoot, Primitive } from 'radix-vue'
 import { useStore } from '@/stores'
