+++ apps/web/src/components/CodemirrorEditor/EditorHeader/index.vue
@@ -1,4 +1,5 @@
 <script setup lang="ts">
+import { altSign, ctrlKey, ctrlSign, shiftSign } from '@md/shared/config'
 import {
   ChevronDownIcon,
   Moon,
@@ -7,7 +8,6 @@ import {
   Settings,
   Sun,
 } from 'lucide-vue-next'
-import { altSign, ctrlKey, ctrlSign, shiftSign } from '@/config'
 import { useStore } from '@/stores'
 import { addPrefix, processClipboardContent } from '@/utils'
