+++ apps/web/src/components/CodemirrorEditor/RightSlider.vue
@@ -1,7 +1,5 @@
 <script setup lang="ts">
 import type { Format } from 'vue-pick-colors'
-import { Moon, Sun } from 'lucide-vue-next'
-import PickColors from 'vue-pick-colors'
 import {
   codeBlockThemeOptions,
   colorOptions,
@@ -10,7 +8,9 @@ import {
   legendOptions,
   themeOptions,
   widthOptions,
-} from '@/config'
+} from '@md/shared/config'
+import { Moon, Sun } from 'lucide-vue-next'
+import PickColors from 'vue-pick-colors'
 import { useDisplayStore, useStore } from '@/stores'
 
 const store = useStore()
