+++ apps/web/src/utils/MDAlert.ts
@@ -1,5 +1,5 @@
+import type { AlertOptions, AlertVariantItem } from '@md/shared/types'
 import type { MarkedExtension, Tokens } from 'marked'
-import type { AlertOptions, AlertVariantItem } from '@/types'
 import { getStyleString } from '.'
 
 /**
