+++ apps/web/src/utils/editor.ts
@@ -1,5 +1,5 @@
 import type CodeMirror from 'codemirror'
-import { altKey, ctrlKey, shiftKey } from '@/config'
+import { altKey, ctrlKey, shiftKey } from '@md/shared/config'
 import { formatDoc } from '@/utils'
 
 interface ToggleFormatOptions {
