+++ apps/web/src/utils/index.ts
@@ -1,7 +1,8 @@
+import type { Block, ExtendedProperties, Inline, RendererAPI, Theme } from '@md/shared/types'
 import type { PropertiesHyphen } from 'csstype'
 import type { ReadTimeResults } from 'reading-time'
-import type { Block, ExtendedProperties, Inline, Theme } from '@/types'
-import type { RendererAPI } from '@/types/renderer-types'
+import { prefix } from '@md/shared/config'
+import { addSpacingToMarkdown } from '@md/shared/utils'
 import DOMPurify from 'isomorphic-dompurify'
 import juice from 'juice'
 import { Marked, marked } from 'marked'
@@ -10,8 +11,6 @@ import * as prettierPluginEstree from 'prettier/plugins/estree'
 import * as prettierPluginMarkdown from 'prettier/plugins/markdown'
 import * as prettierPluginCss from 'prettier/plugins/postcss'
 import { format } from 'prettier/standalone'
-import { prefix } from '@/config/prefix'
-import { addSpacingToMarkdown } from '@/utils/autoSpace'
 import markedAlert from './MDAlert'
 import { MDKatex } from './MDKatex'
