+++ apps/web/src/utils/renderer.ts
@@ -1,8 +1,7 @@
+import type { ExtendedProperties, IOpts, RendererAPI, ThemeStyles } from '@md/shared/types'
 import type { PropertiesHyphen } from 'csstype'
 import type { RendererObject, Tokens } from 'marked'
 import type { ReadTimeResults } from 'reading-time'
-import type { ExtendedProperties, IOpts, ThemeStyles } from '@/types'
-import type { RendererAPI } from '@/types/renderer-types'
 import { cloneDeep, toMerged } from 'es-toolkit'
 import frontMatter from 'front-matter'
 import hljs from 'highlight.js'
