+++ package.json
@@ -42,7 +42,6 @@
     "@ssttevee/multipart-parser": "^0.1.9",
     "@vee-validate/yup": "^4.15.0",
     "@vueuse/core": "^12.5.0",
-    "axios": "^1.8.4",
     "browser-image-compression": "^2.0.2",
     "buffer-from": "^1.1.2",
     "class-variance-authority": "^0.7.1",
@@ -51,7 +50,6 @@
     "core-js": "^3.40.0",
     "cos-js-sdk-v5": "^1.8.7",
     "crypto-js": "^4.2.0",
-    "csstype": "^3.1.3",
     "es-toolkit": "^1.34.1",
     "form-data": "4.0.1",
     "front-matter": "^4.0.2",
@@ -66,15 +64,10 @@
     "pinia": "^3.0.1",
     "qiniu-js": "^3.4.2",
     "radix-vue": "^1.9.12",
-    "reading-time": "^1.5.0",
     "reka-ui": "^2.3.0",
-    "remark-parse": "^11.0.0",
-    "remark-stringify": "^11.0.0",
     "tailwind-merge": "^2.6.0",
     "tailwindcss-animate": "^1.0.7",
     "tiny-oss": "^0.5.1",
-    "unified": "^11.0.5",
-    "unist-util-visit": "^5.0.0",
     "uuid": "^11.1.0",
     "vee-validate": "^4.15.0",
     "vue": "^3.5.13",
@@ -88,9 +81,7 @@
     "@types/buffer-from": "^1.1.3",
     "@types/codemirror": "^5.60.15",
     "@types/crypto-js": "^4.2.2",
-    "@types/mdast": "^4.0.4",
     "@types/node": "^22.13.14",
-    "@types/unist": "^3.0.3",
     "@types/uuid": "^10.0.0",
     "@unocss/eslint-plugin": "^66.3.3",
     "@vitejs/plugin-vue": "^5.2.1",
