+++ packages/shared/src/config/index.ts
@@ -1,0 +1,7 @@
+export * from './ai-services'
+export * from './api'
+export * from './prefix'
+export * from './shortcut-key'
+export * from './store'
+export * from './style'
+export * from './theme'
