+++ packages/shared/src/config/shortcut-key.ts
@@ -1,7 +1,3 @@
-export * from './api'
-export * from './style'
-export * from './theme'
-
 const isMac = /Mac/i.test(navigator.userAgent)
 
 export const ctrlKey = isMac ? `Cmd` : `Ctrl`
