+++ packages/shared/src/config/theme.ts
@@ -1,4 +1,4 @@
-import type { IConfigOption, Theme } from '@/types'
+import type { IConfigOption, Theme } from '../types'
 import { toMerged } from 'es-toolkit'
 
 const defaultTheme: Theme = {
