+++ packages/shared/src/constants/AIConfig.ts
