+++ packages/shared/src/utils/autoSpace.ts
