+++ packages/shared/src/utils/tokenTools.ts
